doc: 2025 Release Schedule
Update the roadmap to include the tentative release schedule for 2025.
This schedule includes the expected versions of Kubernetes and Tekton
used for development and CI testing. The dates lead the overall project
release schedule by two weeks to encourage updates/rebases in
downstream projects, such as the CLI and operator.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>